Trying out a new approach for resource management.
Basically each resource contains an Rc to the context
that created it.  It... actually seems to work very well
so far.  The more intuitive inverse of the context
holding references to the resources takes more work and
is less efficient 'cause it has to be told to do cleanup
by hand.

`tetra` does basically the same thing, looks like it works
quite well.